Ftr: New filtering parameters minSignalBaselineDifference #153
There is a parameter called minSignalBaselineRatio in Peaks Dialog
to filter out peaks on the basis of Signal/Baseline Ratio. Signal
to Baseline ratio of group is calculated from signal/Baseline ratio
of a peak (max of signal baseline ratio of a peak). Signal Basline
Ratio of a peak is peak Intensity divided by max baseline of the peak.

Introduced a new parameter called minSignalBaselineDifference which
is derived from the difference of peak Intensity - max baseline of
a peak.

Issue: #153