Adding calling of pre processing and post processing steps for alignment
To make the new algorithm we would be required to to output the data
required by the python algo for this we call the pre processing steps.
To read the data outputted by the alginment algorithm we would need
process it and read it. Right now we have just added the code in the
function itself. This will be extracted out.